+++ app/com/lynxanalytics/biggraph/frontend_operations/MetaOperations.scala
@@ -20,7 +20,7 @@ class MetaOperations(env: SparkFreeEnvironment) extends OperationRegistry {
   val OtherBoxes = Category("Other boxes", "black", icon = "kraken")
   val AnchorBox = Category("Anchor box", "black", icon = "kraken", visible = false)
 
-  register("Add comment", OtherBoxes)(new DecoratorOperation(_) {
+  register("Comment", OtherBoxes)(new DecoratorOperation(_) {
     override def parameters = List(
       Code("comment", "Comment", language = "plain_text")
     )
@@ -33,14 +33,14 @@ class MetaOperations(env: SparkFreeEnvironment) extends OperationRegistry {
   })
 
   registerOp(
-    "Input box", OtherBoxes,
+    "Input", OtherBoxes,
     List(), List("input"),
     new MinimalOperation(_) {
       override def parameters = List(Param("name", "Name"))
     })
 
   registerOp(
-    "Output box", OtherBoxes,
+    "Output", OtherBoxes,
     List("output"), List(),
     new MinimalOperation(_) {
       override def parameters = List(Param("name", "Name"))