+++ test/com/lynxanalytics/biggraph/frontend_operations/OperationsTestBase.scala
@@ -63,6 +63,12 @@ trait OperationsTestBase extends FunSuite with TestGraphOp {
     lazy val project: RootProjectEditor =
       ctx.allStates(realBox.output("project")).project
 
+    lazy val exportResult: Scalar[String] =
+      ctx.allStates(realBox.output("exportResult")).exportResult
+
+    lazy val table: Table =
+      ctx.allStates(realBox.output("table")).table
+
     def box(operationId: String,
             parameters: Map[String, String] = Map(),
             otherInputs: Seq[TestBox] = Seq(),